sample /tasks/main.yml
sample file in directory [ansible roles/ db/tasks/main.yml] for multi-tier architecture:
Here is an example of the main.yml file in the tasks directory for the db role in a multi-tier architecture:

This file defines the tasks that the db role will perform on the nodes in the database tier. The tasks include installing the database server package, starting and enabling the database service, creating a database, importing a database dump file, waiting for the database to be available, and configuring firewall rules for the database.

Each task uses variables to allow for customization of the role for different database server implementations and environments. For example, the name of the database server package and service can be specified with the db_server_package and db_server_service variables, respectively.

The main.yml file can be used in conjunction with the db role in the playbook to perform the database-related tasks on the nodes in the database tier, providing a way to automate the deployment and configuration of a database server in the multi-tier architecture.