sample tasks/backup.yml
sample file in directory [ansible roles/ db/tasks/backup.yml] for multi-tier architecture:
Here is an example of the backup.yml file in the tasks directory for the db role in a multi-tier architecture:

This file defines the tasks that the db role will perform to backup the database. The tasks include running the database backup command, storing the backup file, and sending a notification about the backup result.

Each task uses variables to allow for customization of the role for different database server implementations and environments. For example, the backup command and destination file can be specified with the backup_command and backup_destination variables, respectively.

The backup.yml file can be used in conjunction with the db role in the playbook to perform database backup tasks, providing a way to automate the backup process for the database server in the multi-tier architecture.