sample /files/my.conf
sample file in directory [ansible roles/ db/files/my.cnf] for multi-tier architecture
Here is an example of the my.cnf file in the files directory for the db role in a multi-tier architecture:

This file is the configuration file for a MySQL database server and provides default values for some of the key configuration options for the database. The file uses variables such as db_port, db_data_dir, and db_socket, to allow for customization of the configuration for different database implementations and environments.

The my.cnf file can be used in the db role to provision the database server with a custom configuration file, providing a way to configure the database server for the specific requirements of the multi-tier architecture.