sample /files/backup_script.sh
sample file in directory [ansible roles/ db/files/backup_script.sh] for multi-tier architecture:
Here is an example of the backup_script.sh file in the files directory for the db role in a multi-tier architecture:

This script is used to create a backup of a MySQL database and is executed on the database server. The script sets up several variables such as DATE, BACKUP_DIR, MYSQL_USER, MYSQL_PASSWORD, MYSQL_HOST, and MYSQL_DB to specify the date of the backup, the location of the backup directory, the username and password of the MySQL user, the hostname of the database server, and the name of the database, respectively.

The script first creates the backup directory if it does not exist, and then uses the mysqldump utility to create a backup of the database. Finally, the script compresses the backup file using gzip.

The script can be used in the db role to provision the database server with a backup script, providing a way to automate the backup process for the database in the multi-tier architecture.